Move eliminateMentionNicknameOrFullName from advanced_text_editor.tsx to text_formatting.tsx
