+++ webapp/channels/src/utils/text_formatting.tsx
@@ -15,6 +15,8 @@ import Constants from './constants';
 import type EmojiMap from './emoji_map.js';
 import * as Emoticons from './emoticons';
 import * as Markdown from './markdown';
+import { displayUsername } from 'mattermost-redux/utils/user_utils';
+import { UserProfile } from '@mattermost/types/users';
 
 const punctuationRegex = /[^\p{L}\d]/u;
 const AT_MENTION_PATTERN = /(?:\B|\b_+)@([a-z0-9.\-_]+)/gi;
@@ -562,6 +564,20 @@ export function getUsernameMentions(text: string): string[] {
     const excludedMentions = new Set(text.match(Constants.SPECIAL_MENTIONS_REGEX) || []);
     const mentions = text.match(AT_MENTION_PATTERN) || [];
     return mentions.filter((mention) => !excludedMentions.has(mention));
+}
+
+export function eliminateMentionNicknameOrFullName(text: string, usersByUsername: Record<string, UserProfile>, teammateNameDisplay: string): string {
+    return getUsernameMentions(text).reduce((msg, mention) => {
+        const username = mention.substring(1);
+        const mentionedUser = usersByUsername[username];
+    
+        if (!mentionedUser) {
+            return msg;
+        }
+    
+        const userDisplayName = displayUsername(mentionedUser, teammateNameDisplay);
+        return msg.replace(`@${username}(${userDisplayName})`, `@${username}`);
+    }, text);
 }
 
 export function autolinkChannelMentions(
